+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 #### 0.2.1 - 2016-06-05
 * Improved structure and formatting of the reference manual.
+* Changed internal stack limit setting mechanism for trampolining.
 * Potentially breaking change: `Stream.Builder.Plus` method was renamed into `Stream.Builder.Combine`.
 * Added `Stream.tryPickJob` and `Stream.tryPickFun`.
 * Fixed `Job.using` to not raise in case given disposable reference is `null`.